requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
shap>=0.43
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14
wbgapi>=1.0
google-cloud-bigquery>=3.0
google-cloud-bigquery-storage>=2.0
rasterio>=1.3
geopandas>=0.14
rasterstats>=0.19
shapely>=2.0
requests>=2.31
openpyxl>=3.1
tqdm>=4.65
